+++ fabric/src/main/java/net/SpectrumFATM/black_archive/fabric/effects/CyberConversionEffect.java
@@ -2,6 +2,7 @@ package net.SpectrumFATM.black_archive.fabric.effects;
 
 import net.SpectrumFATM.black_archive.fabric.BlackArchive;
 import net.SpectrumFATM.black_archive.fabric.entity.ModEntities;
+import net.SpectrumFATM.black_archive.fabric.entity.custom.CybermanEntity;
 import net.SpectrumFATM.black_archive.fabric.entity.custom.DalekPuppetEntity;
 import net.SpectrumFATM.black_archive.fabric.item.ModItems;
 import net.minecraft.entity.LivingEntity;
@@ -10,8 +11,8 @@ import net.minecraft.entity.effect.StatusEffectCategory;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 
-public class DalekNanocloudEffect extends StatusEffect {
-    public DalekNanocloudEffect() {
+public class CyberConversionEffect extends StatusEffect {
+    public CyberConversionEffect() {
         super(StatusEffectCategory.HARMFUL, 0x000000);
     }
 
@@ -27,15 +28,10 @@ public class DalekNanocloudEffect extends StatusEffect {
                 player.damage(player.getDamageSources().generic(), 1.0f);
             }
 
-            if (player.getInventory().contains(new ItemStack(ModItems.DALEK_BRACELET))) {
-                player.removeStatusEffect(BlackArchive.DALEK_NANOCLOUD);
-            }
-
             if (player.getHealth() <= 1f && !player.isDead()) {
-                DalekPuppetEntity dalekSlave = new DalekPuppetEntity(ModEntities.DALEK_PUPPET, player.getWorld());
-                dalekSlave.refreshPositionAndAngles(player.getX(), player.getY(), player.getZ(), player.bodyYaw, player.getPitch());
-                dalekSlave.setPlayerUUID(player.getUuid());
-                player.getWorld().spawnEntity(dalekSlave);
+                CybermanEntity cybermanEntity = new CybermanEntity(ModEntities.CYBERMAN, player.getWorld());
+                cybermanEntity.refreshPositionAndAngles(player.getX(), player.getY(), player.getZ(), player.bodyYaw, player.getPitch());
+                player.getWorld().spawnEntity(cybermanEntity);
                 player.setHealth(0); // Ensure the player is marked as dead
             }
         }